Move groupfinder logic into group service
The groupfinder function currently always loads the group model directly
from the database, this is not ideal when we need to access the same
group multiple times from different parts of the code base within the
same request. This commit adds a new `fetch` method to the group service
and adds a caching layer (similar to the `UserService.fetch`), the
groupfinder then just uses this new fetch method.